test(matrix): hacer que los tests corran en CI, y la ruta de auth en cualquier runner
Cobertura que no corre no es cobertura, y estos no corrían: el job de frontend
solo hacía typecheck. Se añade el paso, simétrico al del backend.

El script `test` del frontend ya existía; lo que le quito es `--passWithNoTests`,
que en un paso de CI convierte "el glob dejó de encontrar ficheros" en verde. Sin
la bandera, cero tests es un fallo, que es lo que debe ser.

Y una separación que estaba mal puesta: `toAlloSession` vivía en `translate.ts`,
entre las tablas de enums, siendo la única traducción que no necesita nada del
binding en tiempo de ejecución — solo la forma de su `Session`, que es un tipo y
desaparece al compilar. Movida a `native/session.ts`, la ruta de autenticación
deja de arrastrar el módulo nativo: `oidcLogin.test.ts` y `session.test.ts` ya no
necesitan doble y corren bajo cualquier runner.

Las otras cuatro suites siguen necesitando jest, y no es un defecto que arreglar:
prueban la traducción de los enums del binding, así que necesitan sus valores en
tiempo de ejecución. Quitárselos significaría redeclarar en Allo los enums del
SDK, que es justo lo que el encargo prohíbe. jest-expo les quita el Flow al
entrypoint de React Native y resuelve el `__mocks__`; el runner propio de bun no
hace ninguna de las dos cosas y muere en `import typeof`. El paso de CI lo dice
por escrito para que nadie lo cambie a `bun test` creyendo que es lo mismo.
