+++ packages/frontend/package.json
@@ -13,7 +13,7 @@
     "build": "expo export --platform web --output-dir dist",
     "build-web": "expo export --platform web --output-dir dist",
     "build-web:prod": "expo export --platform web --output-dir dist",
-    "test": "jest --watchAll=false --passWithNoTests",
+    "test": "jest --watchAll=false",
     "lint": "expo lint",
     "clean": "rm -rf dist .expo"
   },
